chapter_6/digitally_signing/requirements.txt
cffi==1.14.5
pycparser==2.20
PyNaCl==1.4.0
six==1.16.0
